Move python scripts to one directory
