+++ telegraf/python_scripts/raspberry_pi_save_weather_stats/weather.py
@@ -3,7 +3,13 @@
 import sys
 import getopt
 import time
-import random
+
+from bme280 import BME280
+
+try:
+    from smbus2 import SMBus
+except ImportError:
+    from smbus import SMBus
 
 from subprocess import PIPE, Popen, check_output
 import logging
@@ -11,6 +17,7 @@ import requests
 from requests.exceptions import HTTPError, ConnectionError
 
 import platform  # For getting the operating system name
+import subprocess  # For executing a shell command
 
 # TODO Fix logging (after one cycle it wont write in file)
 # create logger with 'spam_application'
@@ -31,6 +38,23 @@ ch.setFormatter(formatter)
 logger.addHandler(fh)
 logger.addHandler(ch)
 
+bus = SMBus(1)
+bme280 = BME280(i2c_dev=bus)
+
+# TODO - this wont be necessary because of telegraf
+# Check for Wi-Fi connection (takhle metoda na Macu nefunguje na raspberry ano)
+def check_wifi():
+    """
+    Checks for wifi connection (works only on Raspberry PI)
+
+    :return: bool True/False depending on connection to wi-fi
+    """
+    if check_output(['hostname', '-I']):
+        return True
+    else:
+        return False
+
+# TODO - this wont be necessary because of telegraf
 def check_connection_with_influxdb(url_domain: str):
     """
     Checks if influxdb is available on http address taken from param
@@ -50,7 +74,7 @@ def check_connection_with_influxdb(url_domain: str):
     response.raise_for_status()
     logger.info(f'Connection to {url_domain} successful')
 
-
+# TODO - redo this part for insertion via telegraf
 def insert_into_influxdb(url_domain: str, database_name: str, data_string: str):
     """
     Runs insert operation in influxdb using curl command
@@ -78,6 +102,25 @@ def get_cpu_temperature() -> float:
     process = Popen(['vcgencmd', 'measure_temp'], stdout=PIPE, universal_newlines=True)
     output, _error = process.communicate()
     return float(output[output.index('=') + 1:output.rindex("'")])
+
+
+def get_compensated_temperature() -> float:
+    """
+    Temporary method compensating heat from CPU
+
+    :return: Float compensated temperature value
+    """
+    comp_factor = 2.25
+    cpu_temp = get_cpu_temperature()
+    raw_temp = bme280.get_temperature()
+    comp_temp = raw_temp - ((cpu_temp - raw_temp) / comp_factor)
+    logger.debug("""
+            comp_factor: {}
+            cpu_temp: {:05.2f} *C
+            raw_temp: {:05.2f} *C
+            comp_temp: {:05.2f} *C
+            """.format(comp_factor, cpu_temp, raw_temp, comp_temp))
+    return comp_temp
 
 
 def print_help():
@@ -114,9 +157,9 @@ def main(argv):
 
     while True:
         try:
-            temperature = random.randint(-20, 40)
-            pressure = random.randint(600, 900)
-            humidity = random.randint(0, 100)
+            temperature = get_compensated_temperature()
+            pressure = bme280.get_pressure()
+            humidity = bme280.get_humidity()
             logger.info("""
             Compensated_Temperature: {:05.2f} *C
             Pressure: {:05.2f} hPa
@@ -126,9 +169,12 @@ def main(argv):
             temperature value={}
             pressure value={}
             humidity value={}""".format(temperature, pressure, humidity)
-            check_connection_with_influxdb(url_domain_arg)
-            insert_into_influxdb(url_domain_arg, database_name_arg, data_string)
-            # TODO Add saving to local file
+            if check_wifi():
+                logger.debug("Connected to wifi")
+                check_connection_with_influxdb(url_domain_arg)
+                insert_into_influxdb(url_domain_arg, database_name_arg, data_string)
+            else:
+                logger.info("No connection to wi-fi")
         except HTTPError as http_err:
             logger.error(f'HTTP error occurred: {http_err}')
         except ConnectionError as conn_err:
