+++ telegraf/python_scripts/save_osx_stats/osx_cpu_fan_stats.py
