+++ telegraf/telegraf_local_PC_stats.conf
@@ -45,6 +45,8 @@
   debug = true
   ## Run telegraf in quiet mode (error log messages only).
   quiet = false
+  
+  # FIXME - This is hardcoded to repository
   ## Specify the log file name. The empty string means to log to stderr.
   logfile = "/Users/michalbudik/Docker_Projekty/sensor_monitor/logs/telegraf.log"
 
@@ -63,7 +65,9 @@
   urls = ["https://eu-central-1-1.aws.cloud2.influxdata.com"]
 
   ## Token for authentication.
-  token = "$INFLUX_TOKEN"
+  # FIXME - Find a way how to save token to file and paste it to telegraf conf
+  #token = "$INFLUX_TOKEN"
+  token = "${INFLUX_TOKEN}"
 
   ## Organization is the name of the organization you wish to write to; must exist.
   organization = "swiftblade1982@gmail.com"
@@ -100,6 +104,7 @@
 [[inputs.system]]
 
 [[inputs.exec]]
+  # FIXME - This is hardcoded to repository
   commands = ["/Users/michalbudik/Docker_Projekty/sensor_monitor/telegraf/save_pc_stats/osx_cpu_fan_stats.py"]
   data_format = "influx"
